tor-proto: Handle incoming CREATE2 with ntor-v3 handshakes

See merge request tpo/core/arti!4176

+++ crates/tor-proto/src/client/reactor/conflux.rs
@@ -386,7 +386,7 @@ impl ConfluxSet {
 
                 // If data is in progress on the leg (inflight > cc_sendme_inc),
                 // then all legs must be closed
-                if inflight >= cwnd.params().sendme_inc() {
+                if inflight >= u32::from(cwnd.params().sendme_inc()) {
                     return Err(ReactorError::Shutdown);
                 }
